feat(pso-algorithm): create test class
