+++ pso-algorithm/modules/__init__.py
@@ -1,0 +1,2 @@
+from modules.particles import *
+from modules.swarm import *
